calculation working hours and check if rest time is valid

+++ src/main/java/com/marcos/punchclock/util/DateUtil.java
@@ -31,15 +31,15 @@ public class DateUtil {
 	
 	public static double getIntervalInHours(Date date1, Date date2) {
 		
-		long intervalInMiliseconds = getIntervalInMiliseconds(date1, date2);
+		double intervalInMiliseconds = getIntervalInMiliseconds(date1, date2);
 		
-		long hourInMiliseconds = 60*60*1000;
+		double hourInMiliseconds = 60*60*1000;
 		
 		return intervalInMiliseconds/hourInMiliseconds;
 		
 	}
 	
-	public static long getIntervalInMiliseconds(Date date1, Date date2) {
+	public static double getIntervalInMiliseconds(Date date1, Date date2) {
 		
 		return date2.getTime() - date1.getTime();
